refactor: convert to tsx
